Speed up nightly testing
Some of our configurations are taking over 24 hours and others are
running long on the machines that we do paratesting on. Speed up these
jobs by either using paratest or increasing the amount of paratest
oversubscription we do. This should get whitebox, linux32, and
llvm-system jobs under 24 hours and get the gasnet tests finishing
faster on our development machines.