Reduce fragmentation for large allocations
jemalloc uses arenas to reduce allocation contention and improve the
performance of concurrent allocations. Typically, a thread will gain
affinity for a specific arena, and future allocations will be satisfied
by that arena. This is great for performance, but it can introduce
fragmentation, which can be a problem for large allocations. To reduce
fragmentation, satisfy large allocations (32 MiB or larger) out of a
single arena. This limits concurrency for large allocations, but those
won't be as common and we're typically not doing them in parallel.

As an example of where things were going wrong -- in a program that
allocates large arrays with multiple trials, we might end up allocating
arrays from thread/arena-0 for the first trial, but then out of
thread/arena-1 for the second trial. When using a fixed heap, jemalloc
had no way to free memory from a different arena so we were seeing out
of memory issues. This isn't as big of an issue when there's no shared
heap because you can allocate more memory from the system or coalesce
memory from a different arena. With a dynamic heap, large allocations
are done through a different interface so this also wasn't as big of a
deal. This was mostly a problem for configurations with a fixed heap
(gasnet with segment large/fast, some ofi configs.)

This was the cause of out-of-memory errors with Arkouda. Arkouda will
create many large arrays across trials and across different client
programs and they often get allocated from different threads. This
caused severe fragmentation and was resulting in OOMs for our nightly
testing. With this issue fixed we can remove some ineffective
workarounds we tried to limit the number of trials and increase the heap
size.

This particular implementation is a bit of a hack, but it should be
effective. We plan to do this in a more principled way using the
extended API to specify which arena to allocate from instead of changing
the arena for the current thread.