Extend the import in some cases and use an explicit use in others for Sort
These submodules required access to chpl_check_comparator and chpl_compare at
least, functions that are defined by the parent module.  We didn't notice the
issue earlier due to user programs having a use of the parent module, allowing
the functions to be found via the point of instantiation.  When an import was
used instead, we were no longer able to resolve those function calls.