Add a test to lock in the corrected behavior
This is a copy of the Sparse portion of correctness.chpl, alterred so that
the use of LinearAlgebra is instead an import and any needed calls from that
are made using qualified access.  With that alteration, this test failed to
compile before the previous commit, due to the Sparse submodule failing to
use its parent module.  The use statement version passed because we were able
to find the needed symbols via the point of instantiation.

This test now works