Clean up cullOverReferences pass (#15967)
Clean up cullOverReferences pass (#15967)

This PR implements a "light" refactor of the cullOverReferences
compiler pass. As part of my work to resolve #14902 I've had to study
the code in cullOverReferences. I found it hard to understand the
structure of the pass because it was composed of a "monolithic"
block of code.

Changes made:

  - Add a context object used to store the state of the pass.
  - Move each step of the pass into a method on the context.
  - Split the body of the 600~ line for-loop iterating over collected
    symbols into separate functions, each prefaced with `check` for
    consistency.
  - Reduce nesting in `check` methods by inverting non-cast if-stmts
    over AST to exit early (e.g. `if (call->isPrimitive(...)) { ...`
    becomes `if (!call->isPrimitive(...)) return false;`).
  - Move and reformat existing comments as necessary.

Add a few more futures illustrating tuple constness issues. These
futures have bogus .good files because the expected output is to be
determined.

Testing:

- [x] ALL on linux64 when CHPL_COMM=none
- [x] ALL on linux64 when CHPL_COMM=gasnet

---

Reviewed by @mppf. Thanks!