Small CQ/counter cleanup.
Reverse a test of the number of counters the fabric domains supports, so
that we do it the same way in both places and I don't get confused
again.

Add SENDs to the operations the AM handler's tx ctx CQ or counter should
be bound for.  (How did we get away with this until now??)  And while
we're doing that, simplify the code by adding READ binding for the AM
handler also.  We don't do READs there now but it doesn't hurt to enable
this and if we ever start doing them we'll need it anyway.