Disable completion counters for now.
I'm having trouble with completion counters with some providers, and
most notably there seems to be some variation in the requirement that tx
endpoints must have CQs no matter what, so that they can gather errors
related to transmits.  So until I can rework the counter logic to allow
for both CQs and counters on an endpoint, and maybe until I see some
evidence that counters make much performance difference, disable support
for them in the comm layer.  (The logic is still there, we just won't
ever choose it.)