Quiet python interop warnings about unrecognized command line option
We started throwing `-Wno-alloc-size-larger-than` for gcc 8+ in 15999.
Python interop testing was warning that this was unrecognized while
cythonizing things. This is because we're using gcc 8 for our compiles
so our makefiles add the option and pass it to cython, but cython then
goes and uses the system gcc 4 version, which doesn't support the flag.
Just filter the warning out like was being done for similar clang
warnings.