Disable restrict warnings for gcc
We're getting some bogus Wrestrict warnings from gcc for memcpy calls in
the string module. We quieted some of these warnings before in 15791,
but now we're seeing them more often. These are false positives and
they're getting noisier so disable the warnings. In general recent
versions of gcc seem to have a much higher rate of false-positives for
new warnings, so I have been more on board with just quieting ones that
don't seem that valuable anyways.