Inline real2int and object2int
The body of these functions is trivial and they can be used in
performance critical code, so inline them. This was noticed while doing
performance tuning of our hash table. I didn't see any concrete
improvement from this, but the body is so small that it makes sense to
do on principle and in case it matters for different compilers/systems.

While here, switch type-punning in real2int from a union to mempcy to
match our other type-punning code. Also move chpl_get_argument_i to
chpltypes.h since it was the last thing in chpltypes.c