Add a new enum to ResolveScope to track progress of resolving imports and uses
This enum has three states:
- not started
- in progress
- completed

When the ResolveScope instance for a scope is created, we traverse the scope to
see if there are ImportStmts or UseStmts present or an already created module
use list.  If so, we mark the scope as not started, and if not, we mark the
scope as already completed.

Update the scopeResolve methods on UseStmt and ImportStmt to be re-entrant
without causing us to fail with an internal assert, as we will now potentially
call those methods multiple times (but only need to resolve the statement once).

Ensure that when we extend a scope with the import or use, we update its
progress.

During lookupForImport, if the scope we would search next for reexported symbols
has not started looking at its own uses and imports, we traverse them (avoiding
this if it is currently in progress, to avoid circularity issues).  We also
now look at the modules being imported themselves prior to the symbols for
unqualified access, as we do for direct module calls later.