Enable parallel array init for all types
Previously, we only used parallel array init for POD types because we
had performance issues with arrays-of-arrays. That performance issue was
resolved in 15895, so we can now enable parallel init for all arrays. We
will probably need to retune our heuristic for switching between serial
and parallel init especially for non-POD types, but to start just use
the existing 2MB threshold.