Optimize map -= operation
For the minus-equal operation on two maps, inline the body of the `remove`
function and then hoist the locking operations and calls to
`table.maybeShrinkAfterRemove` out of the loop. This way, when a lot of elements
are removed, locking, unlocking, and shrinking the table only happens once
instead of once per element removed.