Try suggestion from Michael, which reduces the # of times we traverse a scope
Don't traverse the entire scope when creating a ResolveScope wrapper, only mark
it as not started, and update that as we encounter a use or import or finish
traversing the scopes.

Track the scopes that need to be updated, as we can dive into them instead of
exiting from them through the course of the module.