Update for network atomics != none
This updates the .good files for network atomics ==/!= none and
threads the unstable comm counts flag through the table to support
stable testing.  As part of the effort, it also starts to size each
column to a particular width, though currently it's still hard-coded
to 15 for a stable field and size("unstable") for unstable ones.