Rework finding a provider.
Set up fundamental requirements in the hints.  Add message orderings and
try to find a provider.  If that works we're good, otherwise undo the
message orderings, add delivery_complete instead, and try to find a
(non-RxM) provider again.  If that works we're good, otherwise we can't
continue.  This isn't fully tested but it selects tcp;ofi_rxm on our
chapcs cluster, tcp;ofi_rxm on a Cray CS without a fixed heap,
verbs;ofi_rxm on a Cray CS with a fixed heap, and gni on a Cray XC with
a fixed heap.  That's pretty good, so I'm committing it for the weekend
and will test further (efa) on Monday.