Handle bool tests in the compiler
The compiler does a lot of instantiating `_cond_test`, `_cond_invalid`
and `isTrue`. The common case is to call those on a boolean, when
the behavior of these functions is trivial.

This change handles these common cases in the compiler / preFoldNamed,
yielding minor compilation-time speedup.  I added `skip_cond_test` helper
in the compiler to handle both the case where `_cond_test` was eliminated
and the case where it was not.

WHAT ELSE I CONSIDERED

I tried to avoid inserting `_cond_test` altogether, so that the compiler
implements its functionality directly instead. However, there are too
many cases where _cond_test is used, so that would be too much work
to implement this idea. Also we may want to keep `_cond_test` so that
the user can define it on their data types, see #15904.

I tried to adding `pragma "last resort"` on the generic overloads
of these functions. This does not work when it is invoked, for example,
on a range, because then the compiler promotes the `int` overload,
which is non-last-resort, and we do not want promotion here.

I tried removing the `bool` overloads to reduce the number of candidates
that the compiler considers. Alas this broke the cases that use bools of
non-default sizes. Then, I tried to recognize all bool sizes in the
compiler instead of just the default size. However, in some cases the compiler
expects default-sized booleans downstream. To make those work, we can
have the compiler convert any boolean to the default size. I left that
for another day, as it was not obvious that reducing the number of candidates
was worth it w.r.t. compilation speedup.

GENERIC / NON-GENERIC OVERLOADS

I replaced the generic overloads on `integral` with two concrete overloads
on `int` and `uint`, to reduce the number of generic instantiations.

I collapsed the two generic overloads of `_cond_test` into one,
for the same reason.