Change strategies for marking parallel loop accesses
Previously, we had markVectorizationHazards checking
for patterns we didn't expect to be supported by the vectorizer
we are working with.

The main pattern it is trying to avoid is `alloca` of a temporary
variable for some temp inside the loop. The LLVM code generator
will produce an alloca outside of the loop for this temporary.
Then, the LoopVectorizer will not realize that it should make
the alloca vector-wide and instead will refer to the same temporary
with all of the vector lanes.

The problem with the previous approach is that it was guessing at
which temporary variables would be replaced by LLVM SSA registers.
If a variable is replaced with a register, then the
llvm.loop.parallel_accesses will no longer be relevant for that variable
since llvm.loop.parallel_accesses (and the related llvm.access.group)
are annotating memory instructions.

This commit changes the approach from guessing at which temporary
variables will be replaced. Instead, it identifies pointers that
cannot point to stack memory within the loop, and marks loads/stores
for these with llvm.loop.parallel_accesses / llvm.access.group.
Loads/stores to local variables declared within the loop will not
be so marked and so are not eligible for vectorization in the event
that they are not replaced with SSA registers.

This commit also adds tests that the llvm.loop.parallel_accesses /
llvm.access.group metadata is being generated correctly.