Use conservative order independent hinting for iterators
Yielding loops in follower/standalone iterators are no
longer marked order independent. Instead, these are expected
to be marked by the user or an earlier part of compilation.

Now when an iterator is inlined, the order-independent-ness
of the resulting loop is the && combination of:
 * loop invoking the iterator being order independent
 * loop in invoked iterator being order independent

Additionally, set order-independent for follower loops built by compiler

Comes up for

  forall (i) in zip(1..100) { }

for one reason or another.