Add compiler warning for map.getValue and add docstring (#16072)
Add compiler warning for map.getValue and add docstring (#16072)

Adjust `map.getValue()` so that it uses an embedded compiler error
instead of a where clause. The error fires if `map.getValue()` is
called when the map value type is not a non-nilable class.

Add another `compilerError` that fires when `map.getValue()` is
called and the map value type is an owned class. The error suggests
that the user call `map.getBorrowed()` instead.

Add a docstring describing the behavior and constraints of the
`map.getValue()` method.

Add three tests locking in the behavior of `map.getValue()`.

---

Testing:

- [x] ALL on linux64 when CHPL_COMM=none
- [x] library/standard/Map on darwin when CHPL_COMM=none, -futures
- [x] Updated docs rendered locally

---

Reviewed by @daviditen. Thanks!