Update Crypto module to avoid 2 problems with --llvm
First, on some versions of libssl, OpenSSL_add_all_digests
is a macro. So call it from a `static inline` function that
is defined in CryptoHandlers/openssl_c_support.h.

Second, after PR #16057, the compiler began (correctly)
showing errors for the declarations like

 extern type EVP_MD_PTR = c_ptr(EVP_MD);

In fact EVP_MD_PTR is not a type defined in C at all (at least
with the version of libssl I have been using). So change these
to

 type EVP_MD_PTR = c_ptr(EVP_MD);

and move these and related types to before they are used in the
extern proc declarations.