Fix jenkins processing of Arkouda `make check` errors
Previously, an error in Arkouda's `make check` wasn't getting detected
as an error in jenkins, though it did show up in our nightly emails.
This is because we were marking the test end before emitting the fatal
error, which messed up our junit processing. To fix this, print the
error before marking the test end. This is done by keeping track of the
current test and emitting the test_end in log_fatal_error before
exiting.