Delivery-complete doesn't really apply to receives.
We've been specifying completion-related op_flags for the receive
endpoint attributes as well as the transmit ones, but they don't
actually apply there.  Here, stop doing that.  As a beneficial
side effect, now we select the high-performance efa provider when
we're on AWS and the EFA network is being used.