Add hdf5 tools to PATH for arkouda testing
https://github.com/mhmerrill/arkouda/pull/411 started requiring h5ls for
some I/O tests. Ensure it's available in the PATH when our start_test
infrastructure is the thing installing the HDF5 dependencies. Follow on
to https://github.com/chapel-lang/chapel/pull/16066, which updated PATH
for our nightly testing, which uses a prebuilt HDF5, but missed updating
for users running themselves.