Add a test to lock in that TomlParser has the appropriate uses
This test fails as of this commit because it requires more than just Source from
the TomlReader sibling module (which is used by its parent module), at least for
the parseLoop method on the Parser type.  Now that we don't search the use
statements of higher scopes, it needs an explicit use of TomlReader itself.  We
didn't notice before due to only calling parseLoop from the parent module (and
thus being saved by the point of instantiation).