Fix test of masonTest() that broke in PR #16098
This fixes changes to masonTest() that were introduced in PR #16098
when they assumed that 'args' was an array when the Launcher_test.chpl
test passes it a list.  Here, my fix was to use .indices rather than
.domain to query the argument's size; using an unbounded range to then
slice the list seems like it ought to be supported, yet is not (and
generates a not user-friendly error message), so I'm filing a future
against that as well.