Replace segfault when using qualified access to make a fcf with error message
The code was expecting a qualified access call to be within another CallExpr,
but this is not always the case for fcf accesses.  The initial transformation I
tried ended up causing the function to be executed instead of treated as a fcf,
so leave an error message for now instead of taking the more involved path to
fix what is very much a low priority issue.