Slightly improve support for scalable memory registration.
The primary effect of the work here is to stop communicating the memory
registration keys all-to-all when we do scalable registration, because
in that case we specify the key ourself and thus already know what it
is.  It also slightly accelerates the control flow for getting local
descriptors and remote keys with scalable registration.  Finally, while
here I made some adjustments to the code that decides between basic and
scalable registration.