Adjust module commands so they work with both Tcl and Lmod.
Various build and test environments important to us are converting from
the old Tcl `module` command to the new Lmod-based one.  Unfortunately
the syntax is a bit different: the Lmod versions don't support the '-l'
(dash-ell) option for "long" output, and their '-t' option for terse
output must be applied the command itself rather than to the subcommand.
Here, adapt.  Get rid of the dash-ell options; we'll just live without
the timestamps they gave us.  And migrate the '-t' options from the
subcommand to the command, because as luck would have it, the Tcl-based
stuff supports that syntax also.