Fix a bug with accidentally matching to symbols in scope for a module call
While the transformation of explicit module calls during scope resolution was
appropriate, we didn't consider that it could have occurred at the time when we
resolve unresolved symbols - we didn't check to see if the symbol to be resolved
was used for a call, and as a result didn't think to check if such calls had a
module argument, so were treating them as unqualified calls and thus when a
non-function symbol was in scope and closer than the module's function, we would
match to that (and sometimes get a use-before-def error as a result).

This change adds the necessary check, which resolves the futures tracking the
failure.