Don't set LIBFABRIC_DIR in the Shasta module.
The system libfabric module for Shasta appears to have matured to the
point that we can depend on it to set PKG_CONFIG_PATH so that we can
always find libfabric.  Therefore, stop setting LIBFABRIC_DIR, which we
were using to compensate for not having PKG_CONFIG_PATH.

(As a side effect, this also solves a problem on some Shasta systems
where libfabric.so is in a .../lib64 directory but our util/chplenv
scripts were searching $LIBFABRIC_DIR/lib.  We could not have fixed this
in master and cherry-picked back to release/1.22, with which we're doing
the Shasta module builds, because the recent bundled-libfabric work has
made master's libfabric-related util/chplenv scripts rather different
than the corresponding ones in release/1.22.)