Get ugni hugepage check working with Lmod
Our check to make sure we get a warning if hugepages aren't loaded is
currently failing on systems with Lmod. The problem is that we're
falling back to Tcl still, but we do that in our bashrc after Lmod is
loaded. When we do a qsub a new Lmod session starts, which ends up
reloading hugepages, so we weren't getting our warning. This shouldn't
be a problem once we're no longer failing back to Tcl, but until then
this should get testing working.