Use iteratorGroup in resolveForallHeader()
resolveForallHeader() used to resolve a parallel iterator call
to find the standalone or leader iterator. If a relevant iteratorGroup
were available, resolveForallHeader() used to assert that the iterator
produced by call resolution matched the one stored in the iteratorGroup.

I wrote the code this way to ensure that iteratorGroup worked properly.
Now that the assertions have been holding for many months, I am ready
to use the iterator from the iteratorGroup instead of call resolution,
when the iteratorGroup is available.

I am keeping the code path that uses call resolution because an iteratorGroup
may not be available, ex. when the iterable expression is a variable.

While there: remove an inappropriate assertion in ForallStmt::verify(),
which ensures that the forall body does not have a 'use' statement.
Now we can compiler arkouda with --verify.