Add a primitive "get visible symbols" that prints visible symbols
Add a primitive that prints the symbols visible from the point where it is
called. The primitive is handled near the end of scopeResolve and is then
removed.

To build the list of visible symbols, walk over all gVarSymbols and try to
`lookup()` them from the primitive call site. Build a map from filename to set
of symbols and add the visible ones to the set. Then walk the map and its sets
to print all visible symbols grouped by filename.