Fix a bug in cullOverReferences that caused invalid memory reads (#16268)
Fix a bug in cullOverReferences that caused invalid memory reads
(#16268)

This PR adjusts the pass `cullOverReferences`. It fixes a logic
error that caused the compiler to read invalid memory.

In `CullRefCtx::checkAccessorLikeCall`, the LHS of an "accessor-like"
call (methods marked with `FLAG_REF_TO_CONST_WHEN_CONST_THIS` such
as `_tuple.this`, when the argument is a receiver) will add the
call's receiver as a dependency of the LHS when the LHS is a
reference.

When creating the node to represent the LHS in the dependency graph,
the code incorrectly passed the field index of the receiver. The
type of the LHS is _almost certainly_ not the same as the receiver,
so we cannot carry along the receiver's field index. For now, adjust
the code to care about the constness of the entire LHS instead of a
specific field.

This code may need to be adjusted to be aware of LHS field indices
in the future (e.g. if expanded ref tuples are ever introduced and
one unpacks one ref tuple into another).

---

Testing:

- [x] ALL on linux64 when CHPL_COMM=none
- [x] library/standard/DateTime/testDatetimeTZ.chpl with valgrind
- [x] library/standard/DateTime/testDatetime.chpl with valgrind

---

Reviewed by @mppf. Thanks!