Make nightly compiler and runtime build failures fatal
Previously, if the compiler or runtime build failed we would retry
without optimizations. This was added a long time ago when the build was
less stable and we didn't have smoke testing like we do today. The
original intention was to allow us to get some nightly results even if
the optimized build failed, but I can't remember this happening in many
years. The retry build has confused us recently for transient build
issues, so just make build issues always a fatal error to avoid
confusion. We may lose a night of testing for these configurations with
transient errors, but I don't think that's a huge deal and we should
work to address those issues instead of relying on the retry mechanism.