Make it a fatal error if the llvm runtime is needed and fails to build
Making llvm build failures non-fatal was done in 9e813328f0 so that:

"""
If you try to chpl --llvm and the runtime with clang-included is not
built, you do get a decent error from the compiler. But some
LLVM-based features will still work (extern blocks) if the
clang-included runtime does not build.
"""

However, that was 5+ years ago before we were considering making llvm
the default. Nowadays I think we want to get these failures as early
as we can and limit confusion at `chpl` invocation time.