Track the emitted code size in arkouda perf testing
Track how many statements are generated for Arkouda. Most of Arkouda
compilation time is spent in makeBinary, which tends to be correlated
with the amount of generated code. Start tracking how much we're
generating so that we can the difference across configurations and track
the improvements from some upcoming changes.

The actual implementation is a little ugly since we `tee` the make
output to a file and then grep it. This would be a lot cleaner if we had
support for a `--print-emitted-code-size-file` option (like we have for
`--print-passes-file`)