Quiet messages from lmod when submitting jobs in launchcmd
Lmod prints a lot of information about what modules are being changed or
reloaded, and we see these when doing a qsub. These messages will get
captured in our test output, which we don't want so quiet messages from
lmod to avoid that. Previously, we were setting this in a .bashrc file,
but here we're moving it to launchcmd so it doesn't have to be set
manually by developers.