Align spec/generics compilerDiagnostics example source and output
The compilerDiagnostics.chpl example in spec/generics.rst had drifted
away from the rendered output.

The actual (and checked) output includes the width part of int(64) and
real(64), so include those in the rendered output.

The rendered output includes the types of the args in the 1-argument
warning strings.  Add those to the example code, and the checked
output.