Stop targeting *cronmail*@cray.com mailing addresses
This, along with a pair of PRs on our internal repo, changes our test
structure to stop mailing the *cronmail*@cray.com addresses.  The most
important / failure mails are sent to our emerging Discourse site.
The "all" results are still sent to the SourceForge list (because it
seemed easier at this point than refactoring that logic to not send
anywhere at all).