Remove vestiges of target platform cross compilation support
We used to experimentally support arm cross-compilation by specifying
`CHPL_TARGET_PLATFORM=aarch64`, but that was for a prototype system.
Most of that was removed in 11905, but this was left. I don't think it
has any purpose anymore, so remove it.

If we support a general cross compiling mechanism in the future I
suspect it will be with `CHPL_ARCH`, not `CHPL_PLATFORM` -- 11903