Use CHPL_TARGET_ARCH instead of uname -m to specify 3p --host
When we're cross-compiling for Cray's in our third-party scripts we
throw `--host` in order to tell configure what type of system the code
will be executing on (note that configure `--host` is backwards from our
notion of host vs. target.)

Previously, we were just using `uname -m`, which gives you the host
architecture and on crays the login and compute nodes have to be the
same architecture so this was safe. However, since that work was done we
added a `CHPL_{HOST,TARGET}_ARCH` variable so we can just use that
instead of computing uname. And here we're switching from using the host
arch to the target arch since it's what configure is looking for. In the
cray case they currently have to be the same, but that might not be true
in the future.