Increase the timeout for gasnet-ibv-fast performance testing
When the machine we run on has been up long enough, tests that use a
large fraction of memory like ISx start taking longer and hitting our
timeout under gasnet-ibf-fast. This is a performance issue and not a
hang or correctness issue, so bump the timeout. We'll still track the
issue as a performance regression in our graphs, it just won't create
noise in nightly regression mails.

See https://github.com/Cray/chapel-private/issues/443 for more info